requirements.txt
Flask
PyPDF2
python-docx
spacy
google-generativeai
python-dotenv